+++ ttnn/cpp/ttnn/operations/normalization/CMakeLists.txt
@@ -62,11 +62,10 @@ target_sources(
         layernorm/device/layernorm_op_multi_core.cpp
         layernorm/device/layernorm_op_multi_core_sharded.cpp
         layernorm/layernorm.cpp
-        layernorm_distributed/device/layernorm_post_all_gather_op.cpp
+        layernorm_distributed/device/layernorm_post_all_gather_device_operation.cpp
+        layernorm_distributed/device/layernorm_post_all_gather_program_factory.cpp
         layernorm_distributed/device/layernorm_pre_all_gather_device_operation.cpp
         layernorm_distributed/device/layernorm_pre_all_gather_program_factory.cpp
-        layernorm_distributed/device/multi_core/layernorm_post_all_gather_op_multi_core.cpp
-        layernorm_distributed/device/multi_core/layernorm_post_all_gather_welford_op_multi_core.cpp
         layernorm_distributed/layernorm_post_all_gather.cpp
         layernorm_distributed/layernorm_pre_all_gather.cpp
         rmsnorm/rmsnorm.cpp
